Changed to add function that mimics plugin.spdlogger method
The main difference between this function and plugin.spdlogger is that
since this is not part of an object with a name attribute, the name
is supplied as a function parameter. The logger which gets created and
registered in the global spdlog registry must be accessed by this
string, which is not ideal.

Perhaps this name should be declared as a global const std::string?